+++ src/Client/AndroidUI.h
@@ -199,7 +199,7 @@ public:
         this->setMouseEnabled(true);
         this->setKeypadEnabled(true);
 
-        CCDirector::sharedDirector()->getTouchDispatcher()->addTargetedDelegate(this, -512, true);
+        CCDirector::sharedDirector()->getTouchDispatcher()->addTargetedDelegate(this, -500, true);
 
         this->runAction(CCFadeTo::create(0.5f, 100));